Introduce an "org.kde.kdecoration2" metadata section in JSON file
The idea is to provide further global settings through the metadata
of the decoration plugin. First supported value is "blur" which takes
a boolean value indicating whether the decoration plugin wants the
blur behind effect to be active on it's decoration.

Breeze doesn't need blur, so the key with false would not be needed, but
for a reference plugin it makes sense to provide them all.