Create a BreezeConfig.cmake
Summary:
Breeze creates and installs a BreezeConfig.cmake and
BreezeConfigVersion.cmake.

The idea is to have other Plasma projects which default to Breeze items
(e.g. kstyle in plamsa-integration, kdecoration in KWin) to properly
depend on it.

The content of the BreezeConfig.cmake so far provides whether Breeze
is compiled with support for KDecoration and the plugin id.

This can now be used by e.g. KWin in the following way:
    find_package(Breeze ${PROJECT_VERSION} CONFIG)
    set_package_properties(Breeze PROPERTIES
                           TYPE OPTIONAL
                           PURPOSE "For setting the default window decoration plugin")
    add_feature_info("Breeze-Decoration" BREEZE_WITH_KDECORATION "Default decoration plugin Breeze")

Reviewers: #plasma

Differential Revision: https://phabricator.kde.org/D1343