Show full scrollbar only on mouse over
Summary:
This is a proof of conceptof lighter scrollbars, according to
ideas from the VDG:
show only a tiny placeholder normally, and show the full scrollbar
only on mouse over.

Test Plan:
tested on few applications but i know well the
code is just an early hacky proof of concept,
so i'll redesign it as requested.
being a big visible thing, it will probably need
an option as well to make it behave as it used to.
in action https://www.youtube.com/watch?v=O72qxQHjHcw

Reviewers: #plasma, #breeze, #vdg, hpereiradacosta

Reviewed By: hpereiradacosta

Subscribers: plasma-devel, abetts

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D3131