[Breeze Style] Fix flickering during KPageTreeView animation
QTreeView animates expanding/collapsing of branches. It renders the widget into a temporary pixmap for speed
which is then animated. The pixmap's background, however, is unconditionally filled with the widget's palette's
base color which is different from the window background color onto which we paint when side panel frames are disabled.

This leads to flickering as the background is different during the animation. Depending on the colorscheme
this can be quite noticeable (e.g. Honeycomb).

Differential Revision: https://phabricator.kde.org/D7089