Only include QtQuick support in Breeze KStyle if QtQuick is available
Summary: Currently Breeze hard-depends on QtQuick in order to provide QtQuick-specific theme tweaks, but only for KDE5. This patch decouples KDE4 support from QtQuick support and allows Breeze to be built even when QtQuick is not available (at the loss of the QtQuick specific tweaks of course).

Reviewers: mart

Reviewed By: mart

Subscribers: ngraham, davidedmundson, plasma-devel

Tags: #plasma, #breeze

Differential Revision: https://phabricator.kde.org/D12708