Invert shade button by same logic as keep-above button
Summary:
What do say we invert the shade button when it's in a "checked" state, as we do the keep-above button?

The window decoration theme already inverts the keep-above button's colors when its state is activated, as a "loud" visual cue. The shade button similarly has an active state to indicate, but lacks such a loud cue to help avoid confusion. Depending on a user's other settings, there may be little or no visual indication (other than the nice but modest flip of the button) that a shaded window decoration doesn't belong to the window seen below it.

I'm attaching two screenshots of this patch applied, with a konsole window in both keep-above and shade states, directly over another window's titlebar.

{F6141905}

{F6141930}

Reviewers: #breeze, ngraham, #vdg

Reviewed By: #breeze, ngraham, #vdg

Subscribers: ngraham, plasma-devel

Tags: #plasma, #breeze

Differential Revision: https://phabricator.kde.org/D14389