Change shadow color to 0,0,0
Summary: The current shadow color is Shade Black, which is also used as the ViewBackground color in Breeze dark. This makes the shadow invisible when 2 windows without borders overlap with touching View areas. It also makes color banding worse on cheap displays. Changing the color to black makes the shadow visible on most backgrounds and slightly improves the look of the shadow on cheap displays.

Test Plan:
Before:{F6621844}
After:{F6621843}

Reviewers: #vdg, ngraham, filipf

Reviewed By: #vdg, ngraham, filipf

Subscribers: filipf, ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D19124