Revert "Revert "Reduce the indicator arrow size for press-and-hold menus in QToolButtons""
Summary:
This reverts commit c9aa535863e09bf736b1fb64c182b93530780c5d.

Also makes the small down arrow pixel perfect. I will make the other arrows pixel perfect in a separate patch.

After D19890 (27bcd1be9c2c02067bf15c62fdb7f7e2ad394bd9) was reopened, there was some discussion about the reason it was reverted and the problems it introduced. It is possible that an icon could partially cover this arrow, but it seemed uncommon enough that ignoring the problem was deemed acceptable. The type of control that this arrow is used for is also pretty uncommon because it does not feel very good to use with a mouse or touchpad.

Reviewers: #vdg, #breeze, #plasma, hpereiradacosta, ngraham

Reviewed By: #vdg, #breeze, ngraham

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D26139