Work around an issue with changing the wallpaper's filetype
Summary:
Because of the issue described in T12611, changing the filetype of the Next wallpaper is
problematic and will cause a black screen on upgrade.

This terrible horrible patch works around the issue by using the .png filename extension
even though this is not accurate. That way, the file names are the same as the old
versions of the Next wallpaper and will be loaded correctly. Using the wrong filename
extension does not actually cause any problems, it's just inaccurate. a `README` file is
placed next to the images to explain this.

I'm sorry.

Long-term, we should investigate T12611.

Test Plan:
Have previous Next wallpaper (Ice Cold) installed
Replace it on disk with Volna in original JPEG form
restart plasmashell -> black screen

Try the above again with this patch
restart plasmashell -> new wallpaper visible

Reviewers: davidedmundson, #plasma, ognarb

Reviewed By: ognarb

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D26972