Use "Flow" wallpaper for Plasma 5.19
Summary:
This wallpaper was created by Sandra Smukaste for the Plasma 5.16 wallpaper competition:
https://forum.kde.org/viewtopic.php?f=312&t=160935

And decided upon for Plasma 5.19 in T12724#228410

For the first time, this patch also uses the wallpaper's real name in the UI instead of
"Next". Actual filenames are not adjusted, to ensure that the filenames don't change and
we don't need config changes. See T12705.

Closes T12724
Closes T12705

Test Plan:
1. Use existing "Next" wallpaper
2. Apply this patch, build, and install
3. Restart plasmashell
4. Ta-da: {F8286839}

Reviewers: #vdg, ndavis

Reviewed By: #vdg, ndavis

Subscribers: plasma-devel

Tags: #plasma

Maniphest Tasks: T12724, T12705

Differential Revision: https://phabricator.kde.org/D29449